call digits_to_words.py from bash script

+++ g2p_greek/digits_to_words.py
@@ -350,22 +350,30 @@ def _replace_file_contents(filepath: str):
     move(abs_path, filepath)
 
 
-def convert_kaldi_text(kaldi_text_path):
+def convert_kaldi_text(kaldi_text_path, is_shell_command = False):
     assert os.path.isfile(kaldi_text_path), "Path to the kaldi text file does not exist " \
                                             "or is a directory: {}.".format(kaldi_text_path)
+    new_lines = []
     fh, abs_path = mkstemp()
     with os.fdopen(fh, 'w') as newf:
         with open(kaldi_text_path, "r", encoding="utf-8") as fr:
             line = fr.readline()
             while line:
-                line = " ".join(line.split()[1:]).strip()  # get rid of the utt_id
+                utt_id = line.split()[0]
+                line = " ".join(line.split()[1:]).strip().replace("\n", "")  # get rid of the utt_id
                 new_line = convert_sentence(line)
-                newf.write(new_line)
+                new_line = utt_id + " " + new_line + "\n"
+                if is_shell_command is True:
+                    new_lines.append(new_line)
+                else:
+                    newf.write(new_line)
                 line = fr.readline()
-    # Remove old file
-    os.remove(kaldi_text_path)
-    # Move new file
-    move(abs_path, kaldi_text_path)
+    if is_shell_command is False:
+        # Remove old file
+        os.remove(kaldi_text_path)
+        # Move new file
+        move(abs_path, kaldi_text_path)
+    return new_lines
 
 
 def main():
@@ -392,6 +400,10 @@ def main():
     parser.add_argument("--kaldi-text", required=False, default=None,
                         help="Path to a kaldi text file. This means that the format will "
                              "comply with: utt_id word1 word2 ... wordN")
+    parser.add_argument("--shell-command", action="store_true", dest="is_shell_command",
+                        help="If provided then we will redirect the output to stdout and "
+                             "then you can handle that from your bash script.")
+    parser.set_defaults(is_shell_command=False)
     parser.add_argument("-t", "--test-word", required=False, default="NONE",
                         help="A test word in order to test the functionality of the script.")
     args = parser.parse_args()
@@ -400,8 +412,11 @@ def main():
         print("Converted test word, now exiting...")
         sys.exit(1)
     if args.kaldi_text is not None:
-        convert_kaldi_text(args.kaldi_text)
-        print("Successfully converted kaldi text entries!")
+        out = convert_kaldi_text(args.kaldi_text, args.is_shell_command)
+        if args.is_shell_command is True:
+            print("".join(out))
+        else:
+            print("Successfully converted kaldi text entries!")
         sys.exit(0)
     filepath = args.path
     if filepath == "./files/":
